moved elastic folder under zeeguu_core
