+++ zeeguu_core/elastic/settings.py
